Reduce number of calls to computeSize()
The computeSize() function from litegraph is quite slow (on Firefox at
least) due to measuring text dimensions.  This commit splits the size
update off from toggleWidget() so that it can be called once after
updates are made, instead of being called many times inside a loop.

I suspect this may fix #243, given that having a LoRA Stacker in the
controller adds a significant number of calls to toggleWidget().